add formik, removed smart/dumb pattern
